requirements.txt
google-auth-oauthlib
google-api-python-client
fastapi>=0.103.1
uvicorn>=0.23.2
python-multipart
pydantic>=2.3.0
pydantic-settings>=2.0.3
email-validator
python-dotenv
httpx
passlib[bcrypt]
python-jose[cryptography]
